requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
gunicorn
python-dotenv
psycopg2-binary
schedule